Add initial implementation of webpack loader
